Tweaks implementation related to systemd integration
