+++ lib/puma/launcher.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'sd_notify'
 require 'puma/events'
 require 'puma/detect'
 require 'puma/cluster'
@@ -314,38 +315,18 @@ module Puma
 
     def integrate_with_systemd
       return unless ENV["NOTIFY_SOCKET"]
+      return log "Systemd integration failed. It looks like you're trying to use systemd notify but don't have ruby-sdnotify gem installed" unless defined?(SdNotify)
+
+      require 'puma/systemd'
 
       log "* Enabling systemd notification integration"
-      require "puma/sd_notify"
 
-      @events.on_booted { Puma::SdNotify.ready }
-      @events.on_stopped { Puma::SdNotify.stopping }
+      @events.on_booted { SdNotify.ready }
+      @events.on_stopped { SdNotify.stopping }
 
-      start_watchdog if Puma::SdNotify.watchdog?
-    end
+      systemd = Systemd.new(@events)
 
-    #
-    # Puma's systemd integration allows Puma to inform systemd:
-    #  1. when it has successfully started
-    #  2. when it is starting shutdown
-    #  3. periodically for a liveness check with a watchdog thread
-    #
-
-    def start_watchdog
-      usec = Integer(ENV["WATCHDOG_USEC"])
-      return log "systemd Watchdog too fast: " + usec if usec < 1_000_000
-
-      sec_f = usec / 1_000_000.0
-      # "It is recommended that a daemon sends a keep-alive notification message
-      # to the service manager every half of the time returned here."
-      ping_f = sec_f / 2
-      log "Pinging systemd watchdog every #{ping_f.round(1)} sec"
-      Thread.new do
-        loop do
-          sleep ping_f
-          Puma::SdNotify.watchdog
-        end
-      end
+      systemd.start_watchdog if SdNotify.watchdog?
     end
 
     def spec_for_gem(gem_name)
